+++ types/query/pagination_test.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"testing"
 
+	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
@@ -36,8 +37,8 @@ const (
 )
 
 func TestPagination(t *testing.T) {
-	app, ctx := setupTest()
-	queryHelper := baseapp.NewQueryServerTestHelper(ctx)
+	app, ctx, _ := setupTest()
+	queryHelper := baseapp.NewQueryServerTestHelper(ctx, app.InterfaceRegistry())
 	types.RegisterQueryServer(queryHelper, app.BankKeeper)
 	queryClient := types.NewQueryClient(queryHelper)
 
@@ -59,7 +60,7 @@ func TestPagination(t *testing.T) {
 	res, err := queryClient.AllBalances(gocontext.Background(), request)
 	require.NoError(t, err)
 	require.Equal(t, res.Res.Total, uint64(numBalances))
-	require.Nil(t, res.Res.NextKey)
+	require.NotNil(t, res.Res.NextKey)
 	require.LessOrEqual(t, res.Balances.Len(), defaultLimit)
 
 	t.Log("verify page request with limit > defaultLimit, returns less or equal to `limit` records")
@@ -77,7 +78,7 @@ func TestPagination(t *testing.T) {
 	res, err = queryClient.AllBalances(gocontext.Background(), request)
 	require.NoError(t, err)
 	require.Equal(t, res.Balances.Len(), underLimit)
-	require.Nil(t, res.Res.NextKey)
+	require.NotNil(t, res.Res.NextKey)
 	require.Equal(t, res.Res.Total, uint64(numBalances))
 
 	t.Log("verify paginate with custom limit and countTotal false")
@@ -144,7 +145,7 @@ func TestPagination(t *testing.T) {
 }
 
 func ExamplePaginate() {
-	app, ctx := setupTest()
+	app, ctx, _ := setupTest()
 
 	var balances sdk.Coins
 
@@ -181,10 +182,10 @@ func ExamplePaginate() {
 	}
 	fmt.Println(&types.QueryAllBalancesResponse{Balances: balResult, Res: res})
 	// Output:
-	// balances:<denom:"foo0denom" amount:"100" > res:<total:2 >
+	// balances:<denom:"foo0denom" amount:"100" > res:<next_key:"foo1denom" total:2 >
 }
 
-func setupTest() (*simapp.SimApp, sdk.Context) {
+func setupTest() (*simapp.SimApp, sdk.Context, codec.Marshaler) {
 	app := simapp.Setup(false)
 	ctx := app.BaseApp.NewContext(false, abci.Header{Height: 1})
 	appCodec := app.AppCodec()
@@ -209,5 +210,5 @@ func setupTest() (*simapp.SimApp, sdk.Context) {
 		app.GetSubspace(types.ModuleName), make(map[string]bool),
 	)
 
-	return app, ctx
+	return app, ctx, appCodec
 }