+++ x/ibc/03-connection/client/cli/cli.go
@@ -20,6 +20,7 @@ func GetQueryCmd(clientCtx client.Context) *cobra.Command {
 	ics03ConnectionQueryCmd.AddCommand(flags.GetCommands(
 		GetCmdQueryConnections(clientCtx),
 		GetCmdQueryConnection(clientCtx),
+		GetCmdQueryClientConnections(clientCtx),
 	)...)
 
 	return ics03ConnectionQueryCmd